requirements.txt
openai
langchain==0.0.292
python-dotenv
tiktoken
momento
streamlit